DO Model Deployer for WS Cloud
